// src/enums.ts
export enum REDUCER_TYPES {
  toggleModal = "toggle_modal",
  setModalUser = "set_modal_user"
}

export enum STORIES_REDUCER_TYPES {
  setStories = "set_stories",
  setSingleStory = "set_single_story",
  setNewStoriesBatch = "set_new_story_batch",
  toggleLoading = "toggle_loading",
  startTiming = "start_timing",
}